+++ frontend/src/middleware/normsAbbreviationRedirect.ts
